fix: compare tree-split values in float32, matching XGBoost's internal representation
Found tracing a student-model export mismatch: XGBoost stores/compares
feature values as float32 internally, but this server compared in
float64. A row within float32-rounding distance of a real split
threshold (e.g. 0.001 vs a dumped threshold of 0.00100000005) can take
the wrong branch entirely — not a small probability drift, a flipped
verdict bucket. Previously this exact class of divergence was
misdiagnosed as harmless float32 noise and papered over by loosening
a test tolerance instead of fixing the comparison.

+++ main.go
@@ -45,7 +45,7 @@ func evalTree(n *node, features map[string]float64) float64 {
 	value, ok := features[n.Feature]
 	goLeft := n.DefaultLeft
 	if ok {
-		goLeft = value < n.Threshold
+		goLeft = float32(value) < float32(n.Threshold)
 	}
 	if goLeft {
 		return evalTree(n.Left, features)
